requirements.txt
fastapi
uvicorn
python-dotenv
groq
pypdf
sentence-transformers
faiss-cpu
numpy
requests
streamlit
